requirements.txt
flask
flask_restplus
flask_cors
psycopg2-binary
